backend/requirements-playwright.txt
# Playwright MCP Server Requirements
# Additional dependencies for Playwright functionality

# Core Playwright
playwright==1.40.0
playwright-asyncio==1.40.0

# Browser management and optimization
playwright-chromium==1.40.0
playwright-firefox==1.40.0
playwright-webkit==1.40.0

# Web scraping and data extraction
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0
aiohttp==3.9.1
httpx==0.25.2

# Data processing and storage
pandas==2.1.4
openpyxl==3.1.2
pillow==10.1.0
jsonschema==4.20.0

# Performance and monitoring
psutil==5.9.6
prometheus-client==0.19.0
structlog==23.2.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0

# Task queue and async processing
celery==5.3.4
redis==5.0.1
kombu==5.3.4

# File handling and storage
boto3==1.34.0
minio==7.2.0
dropbox==11.36.0

# Security and encryption
cryptography==41.0.8
pydantic-settings==2.1.0
python-jose==3.3.0
passlib==1.7.4

# Utility libraries
python-multipart==0.0.6
python-dateutil==2.8.2
pytz==2023.3
pyyaml==6.0.1
toml==0.10.2

# HTTP and API clients
fastapi==0.104.1
uvicorn[standard]==0.24.0
aiofiles==23.2.1

# Browser automation extensions
selenium==4.15.2
undetected-chromedriver==3.5.4
chrome-extension-installer==0.6.2

# Image and PDF processing
reportlab==4.0.7
pdfkit==1.0.0
wkhtmltopdf==0.2
img2pdf==0.5.1

# Social media and platform APIs
google-api-python-client==2.108.0
facebook-sdk==3.1.0
twitter-api==1.1.1
linkedin-api==2.2.1

# CAPTCHA solving
2captcha-python==1.2.1
anticaptchaofficial==1.0.36
deathbycaptcha==2.0.3

# Performance monitoring
py-spy==0.3.14
memory-profiler==0.61.0
line-profiler==4.1.1

# Development and testing (will be moved to dev requirements in production)
playwright-test==1.40.0
pytest-playwright==0.4.3
pytest-asyncio==0.21.1
pytest-html==4.1.1
coverage==7.3.2
black==23.11.0
flake8==6.1.0
mypy==1.7.1
isort==5.12.0
pre-commit==3.6.0

# Browser profiling and user agents
fake-useragent==1.4.0
user-agents==2.2.0
cloudscraper==1.2.71

# Proxy support
requests[socks]==2.31.0
pysocks==1.7.1
proxy-rotator==1.0.1

# Cloud storage integration
azure-storage-blob==12.19.0
google-cloud-storage==2.11.0
boto3==1.34.0

# Browser extension development
browser-extension-installer==1.0.0
extension-manager==0.2.0

# Advanced automation features
python-docx==1.1.0
python-pptx==0.6.23
tabulate==0.9.0

# Machine learning for automation optimization
scikit-learn==1.3.2
numpy==1.25.2
tensorflow==2.15.0

# Logging and monitoring
sentry-sdk==1.38.0
loguru==0.7.2
coloredlogs==15.0.1